+++ examples/gridViewExample/src/main/java/io/realm/examples/realmgridview/MyApplication.java
@@ -14,11 +14,17 @@
  * limitations under the License.
  */
 
-package io.realm.internal;
+package io.realm.examples.realmgridview;
 
-class TestUtil {
+import android.app.Application;
 
-    public native static long getMaxExceptionNumber();
-    public native static String getExpectedMessage(long exceptionKind);
-    public native static void testThrowExceptions(long exceptionKind);
+import io.realm.Realm;
+
+public class MyApplication extends Application {
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        Realm.init(this);
+    }
 }