+++ examples/objectServerExample/src/main/java/io/realm/examples/objectserver/CounterActivity.java
@@ -59,7 +59,7 @@ public class CounterActivity extends AppCompatActivity {
         super.onStart();
         if (MyApplication.CURRENT_USER != null) {
             // Create a RealmConfiguration for our user
-            SyncConfiguration config = new SyncConfiguration.Builder(this)
+            SyncConfiguration config = new SyncConfiguration.Builder()
                     .initialData(new Realm.Transaction() {
                         @Override
                         public void execute(Realm realm) {